refactor(application): clean up some unused files
