+++ src/page/Movie/Movie.tsx
@@ -1,10 +1,10 @@
-import Header from '../Header/Header';
+import Header from '../../components/Header/Header';
 import { Grid } from '@material-ui/core';
 import headerBg from '../../assets/images/header-bg.jpg';
 import { makeStyles } from '@material-ui/core/styles';
 import React, { useEffect, FunctionComponent } from 'react';
-import MovieDisplay from '../MovieDisplay/MovieDisplay';
-import CastList from '../CastList/CastList';
+import MovieDisplay from '../../components/MovieDisplay/MovieDisplay';
+import CastList from '../../components/CastList/CastList';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchMovie } from '../../store/movie/thunks';
 import { fetchCast } from '../../store/cast/thunks';
